+++ src/layouts/AppLayout/AppLayout.jsx
@@ -1,18 +1,16 @@
 import { Outlet } from 'react-router-dom';
 import { Footer } from '../../components/features';
-import HeaderLogout from '../../components/features/Header/HeaderLogout/HeaderLogout';
-import HeaderLogin from '../../components/features/Header/HeaderLogin/HeaderLogin';
+import Header from '../../components/features/Header/Header';
 import useAuthStore from '../../stores/auth';
 
 export default function AppLayout() {
   const { isInitialLoading } = useAuthStore();
-  const { auth } = useAuthStore();
 
   return isInitialLoading ? (
     <h1 style={{ height: '100vh' }}>Carregando...</h1> // TODO: add a good looking loading state
   ) : (
     <>
-      {auth ? <HeaderLogin /> : <HeaderLogout />}
+      <Header />
       <Outlet />
       <Footer />
     </>
